Fix multiple 64bit warnings
svn path=/branches/ros-amd64-bringup/; revision=34771

+++ reactos/base/applications/msconfig/msconfig.c
@@ -173,7 +173,7 @@ MsConfigWndProc(HWND hDlg, UINT message, WPARAM wParam, LPARAM lParam)
             break;
 
         case WM_NOTIFY:
-            idctrl = (int)wParam;
+            idctrl = wParam;
             pnmh = (LPNMHDR)lParam;
             if ((pnmh->hwndFrom == hTabWnd) &&
                 (pnmh->idFrom == IDC_TAB) &&
